# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Account Service API",
        "description": "Dokumentasi API untuk Service 1 - Kelola Akun (Tugas 2 IAE)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Ambil daftar semua akun nasabah",
                "operationId": "4bbac367ecf8af0f1c2dbad57ff21dc3",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/accounts/{id}": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Ambil detail & saldo akun tertentu",
                "operationId": "577c024ccebfc85ec3189eb045d43f2b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail akun"
                    },
                    "404": {
                        "description": "Akun tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/accounts/{id}/validation-status": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Cek hasil validasi akun nasabah",
                "operationId": "74d63b396e4dcc6fe300ef9eb69d1f44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengecek status validasi"
                    },
                    "404": {
                        "description": "Akun tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/accounts/{id}/validate": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Validasi akun nasabah dengan orkestrasi SOAP + RabbitMQ",
                "operationId": "0ab7356115ceda6032b8a0022f1433c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Akun berhasil divalidasi, diaudit, dan disebarkan ke broker pesan"
                    },
                    "404": {
                        "description": "Akun tidak ditemukan"
                    },
                    "500": {
                        "description": "Gagal memproses orkestrasi ke server pusat"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/accounts/{id}/validate-enterprise": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "[Tugas 3] Enterprise Validation: Federated SSO + SOAP Audit + RabbitMQ Broadcast",
                "description": "Endpoint ini dilindungi sso.federated middleware yang memverifikasi JWT dari SSO Pusat (iae-sso.virtualfri.id). Setelah token divalidasi dan role dipetakan ke tabel lokal, orkestrasi SOAP audit dan RabbitMQ broadcast dieksekusi penuh. Kirimkan JWT Bearer di header Authorization.",
                "operationId": "5deeab822f8bdfaa48e173e54cbba2f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Akun berhasil divalidasi via Federated SSO + SOAP + RabbitMQ"
                    },
                    "401": {
                        "description": "Unauthorized: JWT SSO tidak valid atau tidak disertakan"
                    },
                    "404": {
                        "description": "Akun tidak ditemukan"
                    },
                    "500": {
                        "description": "Gagal memproses orkestrasi ke server pusat"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Masukkan NIM kamu sebagai API Key (contoh: 102022400263)",
                "name": "X-IAE-KEY",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "description": "[Tugas 3] JWT Bearer Token dari SSO Pusat (iae-sso.virtualfri.id). Diperlukan untuk endpoint /validate-enterprise.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Accounts",
            "description": "Accounts"
        }
    ]
}